version=12.0.RC23
# bug fix: Application rubber-band selection mode not allows atom pick to be selection toggle
# new feature: PLOT propertyX, propertyY, (optional propertyZ)
#   makes a graph for the currently selected atom set
#   TODO -- still working on this -- just preliminary -- bounds/centering proplems
#
#   also equivalent for quaternion and ramachandran commands, but preferred for 
#   plot quaternion...
#   plot ramachandran...
# new feature: mapProperty {atomSetFrom}.property1 {atomSetTo}.property2 (optional propertyKey)
#   Allows mapping of property data from one set of atoms to another 
#   based on a key such as atomno or resno (default atomno). 
#   Property2 must be "settable".
#   For example, creating a quaternion map, then mapping that data
#   with straightness, and then selecting a very specific subset of 
#   the quaternion data and using that selection to select data 
#   in the original model:
#
#  load 3cc2.pdb
#  calculate straightness
#  set quaternionframe "A"
#  quaternion r difference
#  display {rna}
#  mapProperty  {1.1 and *.P}.straightness {2.1 and rna}.property_x resno
#  select 2.1 and rna; color property_x;spacefill 0.1
#  select 2.1 and rna and within(2.0,{1.7700001 -0.49 2.62})
#  mapProperty {2.1 and rna and selected}.selected {1.1 and *.P}.selected resno
# new feature: mapProperty SELECTED {atomSet} [key (default ATOMNO)]
#   same as mapProperty {selected}.selected {atomSet}.selected key
# new feature: quaternionFrame "B" for nucleic C4'/P/C4'
# bug fix: atom deletion can result in array length error
# feature change: quaternionFrame "C" for nucleic acids now 
#                 set as per:
#    Sarver M, Zirbel CL, Stombaugh J, Mokdad A, Leontis NB. 
#    FR3D: finding local and composite recurrent structural motifs in RNA 3D structures. 
#    J. Math. Biol. (2006) 215-252

+++ Jmol/src/org/jmol/script/ScriptMathProcessor.java
@@ -826,14 +826,14 @@ class ScriptMathProcessor {
     float f0 = ScriptVariable.fValue(args[0]);
     float f1 = ScriptVariable.fValue(args[1]);
     float df = ScriptVariable.fValue(args[2]);
-    Float[] data = (isListf ? (Float[]) x1.value : null);
+    float[] data = (isListf ? (float[]) x1.value : null);
     String[] sdata = (isListf ? null : (String[]) x1.value);
     int nbins = (int) ((f1 - f0) / df + 0.01f);
     int[] array = new int[nbins];
     String[] sout = new String[nbins];
     int nPoints = (isListf ? data.length : sdata.length);
     for (int i = 0; i < nPoints; i++) {
-      float v = (isListf ? data[i].floatValue() : Parser.parseFloat(sdata[i]));
+      float v = (isListf ? data[i] : Parser.parseFloat(sdata[i]));
       int bin = (int) ((v - f0) / df);
       if (bin < 0)
         bin = 0;
